Switch to dev emacs instance improvements
We already had <f1> launching a dev version of emacs.

Now we also focus to it so that we switch to the right workspace if
the window manager is so configured.

Also, in the dev instance we define <f1> to be "back to main". This
way we can quickly check if what we're working on works and go back
quickly. Oh, the joy of laziness...

GitOrigin-RevId: 8be8606b4543862c6572a0d4e294116e11e70276