🐛 fix(neo): Fix a number of annoying problems with magit (#56)
- we had a problem w/ after-save hook that caused saved buffers to be
switched to Magit mode. We do now what is recommended in
https://docs.magit.vc/magit/Modes-and-Buffers.html#Automatic-Refreshing-of-Magit-Buffers-1. Never
understood why setting the hook in use-package was a problem. One
noticeable fact is that the recommended way uses (add-hook HOOK FN t)
and maybe the use-package version doesn't have that 't'.

- we added 'workspaces' to the sections in magit status. Also changed
the visibility of some of the sections.

- added a 'b K' shortcut to delete branches that have already been
merged into main. This should be a separate PR, but I was touching
this file and sinned.

GitOrigin-RevId: 6d84b711463ff4cc5b06bd637b7e2b02c579dd0c