✨ feat(neo): Add a way to create and switch to worktrees from magit (#93)
This PR adds a 'w' key (dropping the default binding to magit-am that
we do not use)

The idea is that typing 'w' on an issue creates a branch and a
worktree if they do not exist and switch the projectile project to the
worktree directory. This in turns does the right prespective and
treemacs changes.

Typing 'w' on a worktree simply switch to the workspace.

There're still a few quirks, but we'll treat them as bugs as more we
gain more experience with this new feature.

Closes: #92
Closes: #85
GitOrigin-RevId: c50d974104a33bdc9d402df5412fa3d58496cb73